+++ Transformation/haar_reverse_multilevel.m
@@ -1,5 +1,5 @@
 function haar_matrix = haar_reverse_multilevel(original_matrix, n_transforms)
-    haar_matrix = original_matrix;
+    haar_matrix = double(original_matrix);
     for i = n_transforms-1:-1:0
         rows = 1:(size(haar_matrix,1)/2^i);
         columns = 1:(size(haar_matrix,1)/2^i);
