+++ Transformation/haar_transform_multilevel.m
@@ -1,5 +1,5 @@
 function haar_matrix = haar_transform_multilevel(original_matrix, n_transforms)
-    haar_matrix = original_matrix;
+    haar_matrix = double(original_matrix);
     for i = 1:n_transforms
         rows = 1:size(haar_matrix,1)/i;
         columns = 1:size(haar_matrix,1)/i;
